Added inquirer for CLI
With this commit, the CLI (when called) will read the ~/.aws/credentials to list all the AWS Profiles on the system.
Also a list of regions is provided to choose from, for cognito pool.
All the above choices are list type to choose from and user input is supported with fuzzy search on data.